% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_accidents.R
\docType{data}
\name{hk_accidents}
\alias{hk_accidents}
\title{Dataset on accidents in Hong Kong between 2014 and 2019.}
\format{
A data frame with 95821 rows and 33 variables:
\describe{
\item{Date_Time}{\code{dttm} object containing the date and time of the
incident.}
\item{OBJECTID}{Numeric. Identifier for the row.}
\item{Year}{Numeric. Year in which incident has occurred.}
\item{Serial_No_}{Numeric. Serial number for cross-matching with data
between \code{hk_accidents}, \code{hk_casualties} and \code{hk_vehicles}.}
\item{Severity}{String representing the severity of the incident. Values
include: \code{Fatal}, \code{Serious}, \code{Slight}.
}
\item{District_Council_District}{String. Code representing district in
which the incident has occurred. Join with \code{hkdistrict_summary} to extract
full labels. District boundaries are accurate as of 2019.}
\item{Hit_and_Run}{Logical. Indicates whether the incident is a
hit-and-run.}
\item{Weather}{String. Description of weather conditions.}
\item{Rain}{String. Description of rain conditions.}
\item{Natural_Light}{String. Description of lighting conditions.}
\item{Junction_Control}{String. Description of junction condition.}
\item{Vehicle_Movements}{String. Description of the number of moving
vehicles.}
\item{Type_of_Collision}{String. Description of type of collision.}
\item{No_of_Vehicles_Involved}{Numeric. Number of vehicles involved in
collision.}
\item{No_of_Casualties_Injured}{Numeric. Number of casualties injured in
collision.}
\item{Grid_E}{}
\item{Grid_N}{}
\item{latitude}{Numeric. Latitude values.}
\item{longitude}{Numeric. Longitude values.}
\item{Overtaking}{CURRENTLY UNUSED.}
\item{Within_70m}{Logical.}
\item{Precise_Location}{Character. Text description of the precise location.}
\item{Accident}{Logical. Indicates whether incident was an accident.}
\item{Junction_Type}{Character.}
\item{Crossing_Control}{Character.}
\item{Crossing_Type}{Character.}
\item{Street_Name}{Character.}
\item{Road_Type}{Character.}
\item{Cycle_Type}{Character.}
\item{Type_of_Collision_with_cycle}{Character.}
\item{Structure_Type}{Character.}
\item{Road_Hierarchy}{Character.}
\item{Road_Ownership}{Character.}
...
}
}
\source{
\url{https://www.td.gov.hk/}
}
\usage{
hk_accidents
}
\description{
A dataset containing traffic accidents between 2014 and 2019.
}
\keyword{datasets}
