% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkagepop19.R
\docType{data}
\name{hkagepop19}
\alias{hkagepop19}
\title{Land-based non-institutional population by District Council district and age group}
\format{
A data frame with 18 rows and 8 variables:
\describe{
\item{District_ZH}{District Name in Chinese}
\item{District_EN}{District Name in English}
\item{Age_0_14}{Population aged between 0 and 14}
\item{Age_15_24}{Population aged between 15 and 24}
\item{Age_25_44}{Population aged between 25 and 44}
\item{Age_45_64}{Population aged between 45 and 64}
\item{Age_65}{Population aged 65+}
\item{TotalPopulation}{Total population per district}
...
}
}
\source{
\url{https://www.censtatd.gov.hk/hkstat/sub/sp150.jsp?productCode=FA100096}
}
\usage{
hkagepop19
}
\description{
Land-based non-institutional population by District Council district and age group
}
\details{
This dataset is updated as of 2019. Numbers represent estimates of population
based on calculated census percentages.
}
\keyword{datasets}
