% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkdc.R
\docType{data}
\name{hkdc}
\alias{hkdc}
\title{Dataset with public domain information on Hong Kong District Councillors (elected 2019).}
\format{
A data frame with 452 rows and 29 variables:
\describe{
\item{ConstituencyCode}{Constituency code for the District Council}
\item{Constituency_ZH}{Constituency name in Chinese}
\item{Constituency_EN}{Constituency name in English}
\item{District_ZH}{District name in Chinese}
\item{District_EN}{District name in English}
\item{Region_ZH}{Region name in Chinese}
\item{Region_EN}{Region name in English}
\item{Party_ZH}{Party name in Chinese}
\item{Party_EN}{Party name in English}
\item{DC_ZH}{District Councillor name in Chinese}
\item{DC_EN}{District Councillor name in English}
\item{FacebookURL}{URL of the DC's Facebook Page}
\item{DCPageURL}{URL of the DC's official page}
\item{Address}{Office address of the DC}
\item{Phone}{Phone number of the DC}
\item{Fax}{Fax number of the DC}
\item{Email}{Email address of the DC}
\item{WebsiteURL}{URL of the DC's non government page}
\item{DCProjectPageURL}{URL of the DC's page listing on Hong Kong Districts Info}
\item{ElectionYear}{Year when the DC is elected}
\item{ElectionDate}{Date when the DC is elected}
\item{CandidateNum}{Candidate number of the DC when elected}
\item{Occupation}{Occupation of the DC when elected}
\item{Political}{Political Party of the DC when elected}
\item{Camp}{Political leaning/camp of the DC when elected}
\item{Vote}{Number of votes won}
\item{VotePercentage}{Share of vote}
\item{Gender}{Gender of the DC}
\item{Tag}{Granualar political leaning/camp of the DC when elected}
...
}
}
\source{
\url{https://www.districtcouncils.gov.hk}

\url{https://dce2019.hk01.com/}
}
\usage{
hkdc
}
\description{
A dataset containing public domain information of the 452 Hong Kong District Councillors
elected in 2019.
}
\keyword{datasets}
