% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkdistrict_summary.R
\docType{data}
\name{hkdistrict_summary}
\alias{hkdistrict_summary}
\title{Dataset summarising the labels and regions of Hong Kong's Districts}
\format{
A data frame with 18 rows and 5 variables:
\describe{
\item{Code}{Constituency code for the District}
\item{District_EN}{ }
\item{District_ZH}{ }
\item{Region_ZH}{ }
\item{Abbrev}{ }
...
}
}
\usage{
hkdistrict_summary
}
\description{
This is a summary table that details the region classification and abbreviated labels
of Hong Kong's districts.
}
\keyword{datasets}
