% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkstreetnames20.R
\docType{data}
\name{hkstreetnames20}
\alias{hkstreetnames20}
\title{Dataset with Hong Kong Street Names in 2020.}
\format{
A data frame with 4553 rows and 3 variables:
\describe{
\item{DC}{Constituency code for the District Council}
\item{StreetNames_EN}{Street name in English}
\item{StreetNames_ZH}{Street name in English}
...
}
}
\source{
\url{https://www.landsd.gov.hk}
}
\usage{
hkstreetnames20
}
\description{
A dataset containing street names in Hong Kong as at 2020.
}
\keyword{datasets}
