% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_vehicles.R
\docType{data}
\name{hk_vehicles}
\alias{hk_vehicles}
\title{Dataset of vehicles involved in collisions between 2014 and 2019.}
\format{
A data frame with 153062 rows and 8 variables:
\describe{
\item{OBJECTID}{}
\item{Year}{}
\item{Serial_No_}{}
\item{Driver_Age}{}
\item{Driver_Sex}{}
\item{Year_of_Manufacture}{}
\item{Severity_of_Accident}{}
\item{Vehicle_Class}{}
...
}
}
\source{
\url{https://www.td.gov.hk/}
}
\usage{
hk_vehicles
}
\description{
A dataset containing information of collisions by vehicles between 2014
and 2019.
}
\keyword{datasets}
